aiquery/requirements.txt
# Core dependencies (pinned for reproducibility)
pandas==2.2.3
numpy==2.2.6
qiskit==2.2.0
qiskit-aer==0.17.2

# BigQuery and Google Cloud
google-cloud-bigquery==3.38.0
bigframes==2.21.0

# Optional dependencies for enhanced functionality
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.3.0
folium==0.17.0

# Development and testing
pytest==7.4.3
jupyter==1.1.1
ipykernel==6.30.1
python-dotenv==1.0.1
flake8==7.1.1
ruff==0.8.0
mypy==1.13.0

# Data processing
scikit-learn==1.6.1
scipy==1.15.3
xarray==2024.7.0
netCDF4==1.7.2